+++ packages/shared/components/modals/index.js
@@ -6,5 +6,4 @@ export { default as Modal } from './Modal.svelte'
 export { default as NetworkSummaryModal } from './NetworkSummaryModal.svelte'
 export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
 export { default as ProfileActionsModal } from './ProfileActions.svelte'
-export { default as ProposalDetailsMenu } from './ProposalDetailsMenu.svelte'
 export { default as RecipientAccountSelector } from './RecipientAccountSelector.svelte'
