+++ packages/shared/locales/en.json
@@ -1054,6 +1054,10 @@
             "body": "Please provide the information listed below to add a proposal.",
             "addToAllAccounts": "Add the proposal to all accounts"
         },
+        "editProposal": {
+            "title": "Edit proposal",
+            "body": "Please provide the new url for the node that contains this proposal."
+        },
         "voteForProposal": {
             "title": "Vote for proposal",
             "body": "You're about to vote for {proposal} proposal.",
@@ -1267,6 +1271,7 @@
         "vote": "Vote",
         "addProposal": "Add proposal",
         "removeProposal": "Remove proposal",
+        "changeNode": "Change Node",
         "stopVoting": "Stop voting",
         "revote": "Revote"
     },
